Move bitcoin docker to common trait
