+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod bitcoin;
 pub mod create_comit_app;
 pub mod docker;
 pub mod executable;
